adicionando atualizações

+++ mavenproject1/mercado/src/main/java/br/edu/iff/mercado/entidades/Produto.java
@@ -6,6 +6,7 @@
 package br.edu.iff.mercado.entidades;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Date;
 import javax.persistence.Basic;
@@ -59,9 +60,9 @@ public class Produto implements Serializable {
     @Temporal(TemporalType.TIMESTAMP)
     private Date dtPromocao;
     @Column(name = "vl_unidade")
-    private BigInteger vlUnidade;
+    private BigDecimal vlUnidade;
     @Column(name = "vl_promocao")
-    private BigInteger vlPromocao;
+    private BigDecimal vlPromocao;
     @Size(max = 100)
     @Column(name = "ds_descricao")
     private String dsDescricao;
@@ -114,19 +115,19 @@ public class Produto implements Serializable {
         this.dtPromocao = dtPromocao;
     }
 
-    public BigInteger getVlUnidade() {
+    public BigDecimal getVlUnidade() {
         return vlUnidade;
     }
 
-    public void setVlUnidade(BigInteger vlUnidade) {
+    public void setVlUnidade(BigDecimal vlUnidade) {
         this.vlUnidade = vlUnidade;
     }
 
-    public BigInteger getVlPromocao() {
+    public BigDecimal getVlPromocao() {
         return vlPromocao;
     }
 
-    public void setVlPromocao(BigInteger vlPromocao) {
+    public void setVlPromocao(BigDecimal vlPromocao) {
         this.vlPromocao = vlPromocao;
     }
